Спринт 23.3

+++ app/src/main/java/com/practicum/playlistmaker/playlist/data/PlaylistRepositoryImpl.kt
@@ -15,6 +15,7 @@ import com.practicum.playlistmaker.playlist.domain.model.Playlist
 import com.practicum.playlistmaker.search.domain.model.Track
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.withContext
 import java.io.File
@@ -37,9 +38,14 @@ class PlaylistRepositoryImpl(
         appDatabase.playlistDao().deletePlaylist(convertPlaylistToEntity(playlist))
     }
 
-    override fun getPlaylist(): Flow<List<Playlist>> {
+    override fun getPlaylists(): Flow<List<Playlist>> {
         return appDatabase.playlistDao().getPlaylists()
             .map { entities -> entities.map { entity -> convertEntityToPlaylist(entity) } }
+    }
+
+    override suspend fun getPlaylistById(id: Long): Playlist {
+        val entity = appDatabase.playlistDao().getPlaylistById(id)
+        return convertEntityToPlaylist(entity)
     }
 
     override suspend fun updatePlaylist(playlist: Playlist) {
@@ -81,7 +87,10 @@ class PlaylistRepositoryImpl(
     }
 
     override suspend fun deletePlaylistTrack(track: Track) {
-        appDatabase.playlistTrackDao().deletePlaylistTrack(convertPlaylistTrackToEntity(track))
+        val playlists = getPlaylists().first()
+        if (!playlists.any { it.trackIds.contains(track.trackId) }) {
+            appDatabase.playlistTrackDao().deletePlaylistTrack(convertPlaylistTrackToEntity(track))
+        }
     }
 
     override suspend fun getPlaylistTrack(trackId: String): Track? {
@@ -92,6 +101,16 @@ class PlaylistRepositoryImpl(
     override suspend fun getPlaylistTracks(trackIds: List<String>): List<Track> {
         val entities = appDatabase.playlistTrackDao().getPlaylistTracks(trackIds)
         return entities.map { entity -> convertEntityToPlaylistTrack(entity) }
+    }
+
+    override suspend fun deleteTrackFromPlaylist(
+        track: Track,
+        playlist: Playlist
+    ) {
+        val updatedTrackIds = playlist.trackIds.filter { it != track.trackId }
+        val updatedPlaylist = playlist.copy(trackIds = updatedTrackIds, trackCount = updatedTrackIds.size)
+        updatePlaylist(updatedPlaylist)
+        deletePlaylistTrack(track)
     }
 
     private fun convertEntityToPlaylist(entity: PlaylistEntity): Playlist {
